# openapi/ontology.yaml
openapi: 3.1.0
info:
  title: Ontology OpenAPI
  description: Ontology OpenAPI Specification
  version: 0.0.1
servers:
  - url: http://localhost:4242
    description: Local development server
...
paths:
  "/v1/ontologies":
    get:
      tags:
        - Ontology
      summary: List Ontologies
      operationId: listOntologies
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        "200":
          description: Visible Ontologies
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OntologyPage"
    post:
      tags:
        - Ontology
      summary: Create an Ontology
      operationId: createOntology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateOntologyReq"
      responses:
        "201":
          description: Created Ontology
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Ontology"
...
  /v1/ontologies/{ontologyId}/agents/{agentId}/conversations:
    get:
      tags:
        - Conversation
      summary: List Conversations
      operationId: listConversations
      responses:
        '200':
          description: List Conversations response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
    post:
      tags:
        - Conversation
      summary: Create Conversation
      operationId: createConversation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: Created Conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
    parameters:
      - name: agentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/agents/{agentId}/conversations/{id}:
    get:
      tags:
        - Conversation
      summary: Get Conversation
      operationId: getConversation
      responses:
        '200':
          description: Get Conversation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
    parameters:
      - name: agentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/agents/{agentId}/conversations/{id}/action-confirmations:
    post:
      tags:
        - Conversation
      summary: Confirm Action
      operationId: confirmAction
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmActionRequest'
        required: true
      responses:
        '200':
          description: Confirm Action response
          content:
            application/json:
              schema:
                type: object
    parameters:
      - name: agentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/agents/{agentId}/conversations/{id}/confirm-rule-transform:
    post:
      tags:
        - Conversation
      summary: Confirm Rule Transform
      operationId: confirmRuleTransform
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
        required: true
      responses:
        '200':
          description: Confirm Rule Transform response
          content:
            application/json:
              schema:
                type: object
    parameters:
      - name: agentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OntologyId'
...
components:
  schemas:
    CreateOntologyReq:
      type: object
      additionalProperties: false
      required:
        - name
        - id
      properties:
        id:
          type: string
          maxLength: 32
        name:
          type: string
          minLength: 1
          maxLength: 32
        description:
          type: string
          maxLength: 1024
...
    OntologyPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/Ontology"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    Ontology:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - createdAt
        - updatedAt
...
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ontologyId:
          type: string
        title:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateConversationRequest:
      type: object
      properties:
        title:
          type: string
    ConfirmActionRequest:
      type: object
      properties:
        actionTypeId:
          type: string
          format: uuid
        previewToken:
          type: string
...
    ResourceId:
      type: string
      description: Caller-defined stable ID used in API paths and references
      example: employee
      pattern: ^[A-Za-z][A-Za-z0-9_-]{0,159}$
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
        content:
          type: string
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolTrace'
        createdAt:
          type: string
          format: date-time
...
    ToolTrace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        arguments:
          type: object
          additionalProperties: {}
        result: {}
        mutationPreview:
          type: boolean
...
  parameters:
...
    OntologyId:
      name: ontologyId
      in: path
      description: Caller-defined stable Ontology ID
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
...